requirements.txt
spacy-udpipe
python-dotenv
pyarrow
pandas
alive-progress